chores: code split
